+++ frontend/src/pages/AppointmentProgressPage/AppointmentProgressPage.styles.tsx
@@ -15,18 +15,4 @@ const StyledContainer = styled.div`
   `)}
 `;
 
-const StyledLeftContainer = styled.div`
-  display: flex;
-  width: 45.2rem;
-  flex-direction: column;
-  gap: 1.8rem;
-`;
-
-const StyledRightContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 1.2rem;
-  align-items: center;
-`;
-
-export { StyledContainer, StyledLeftContainer, StyledRightContainer };
+export { StyledContainer };
