+++ frontend/src/pages/AppointmentProgressPage/components/AppointmentProgressInputs/AppointmentProgressInputs.styles.tsx
@@ -1,0 +1,17 @@
+import styled from '@emotion/styled';
+
+const StyledLeftContainer = styled.div`
+  display: flex;
+  width: 45.2rem;
+  flex-direction: column;
+  gap: 1.8rem;
+`;
+
+const StyledRightContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 1.2rem;
+  align-items: center;
+`;
+
+export { StyledLeftContainer, StyledRightContainer };
